Just found out that THREADS_PER_BLOCK is ( 128 ), so falling back to a loop over 'valid' indices rather than a hard-coded summation.
Since there can be cases when N is not a multiple 128, looping over 'valid' indices ensures that we don't do extra work.
Also, zeroing out 'invalid' indices and adding them through a generic (idx->idx+127) sum induces pairs of reads and writes which could be time-consuming compared to a O(1) assignment of maxIdx.